feat: ExplorerFontList 생성 (#39)
